+++ src/model/mod.rs
@@ -1,8 +1,7 @@
 pub(crate) mod config;
-pub(crate) mod model_config;
-pub(crate) mod model_playlist;
+pub(crate) mod playlist;
 pub(crate) mod mapping;
 pub(crate) mod api_proxy;
 pub(crate) mod stats;
 pub(crate) mod xmltv;
-pub(crate) mod model_xtream;
+pub(crate) mod xtream;